+++ assets/js/theme/graphql/GraphQL.js
@@ -167,6 +167,10 @@ export default class GraphQL {
                             }
                         }
                     }
+
+                    brand {
+                        name
+                    }
                 
                     defaultImage {
                         url(width: 500, height: 500)